Clarify readme to include "marshal=True"
I spent a fair amount of time to figure this out, along with a few others who have posted similar questions in the issues. 

It isn't exactly clear that `diff` by default does not provide you with key-names that you can use, so this will help people save some time by knowing that the `marshal` option seems to exist for this purpose.