setup.py: migrate to pyproject.toml
Embrace the PEP5158 standard but keep on using setuptools. Building is done
the same way (python -sBm build) so there are no functional changes to
the build artifacts. The motivation for this change is that switching to
another build system in future is trivial when using PEP518. An added
bonus is this file integrates nicely with linters and formatters as
well.

Signed-off-by: Cory Todd <cory.todd@canonical.com>