packaging: switch back to requirements.txt
The pyproject.toml spec is highly opinionated about installing
dependencies. As such, locally development is unpleasant when running
pip install . since it will always install the package, not just our
depends. This complicates testing "the right changes".

To promote a better experience, return to using requirements files and
make pyproject.toml depend on them dynamically so we don't have to
maintain the information in multiple places.

Signed-off-by: Cory Todd <cory.todd@canonical.com>