introduce Serializer class
We have never handled the JSONDecodeError so consumers are assumed to
be catching that exception on their end. This means we cannot wrap all
the parser errors into a unified type (e.g. ValueError) without breaking
the API. Keep API compatibility on the loader functions by introducing a
Serializer class that the cli tool can utilize. This allows the cli to
avoid knowing which specific serializer errors to handle. The library
itself maintains its current API.

Signed-off-by: Cory Todd <cory.todd@canonical.com>