cli: handle malformed input files
Do not show a stacktrace on the cli tool when an invalid file is
specified. Instead, use the new Serializer class to abstract the load
process and handle the unified ValueError. Also handle FileNotFound
because that's a fairly normal user input scenario as well. Set return
code on sys.exit to indicate abnormal exit.

Fixes #71

Signed-off-by: Cory Todd <cory.todd@canonical.com>