Share _obj_diff results between steps of list diff
The backtracking pass of the list diff algorithm in `_list_diff_0` was
calling `_obj_diff` on pairs of objects that had already been diffed in
the forward pass earlier in the `_list_diff` method. This commit saves
the `_obj_diff` results from the forward pass and reuses them in the
backtracking pass so we don't duplicate the recursive calls.